add: swagger fix & isVerified check in resetPass
