+++ .env-example
@@ -1,3 +1,4 @@
+BASE_URL=
 DATABASE_URL=
 PORT=
