+++ package.json
@@ -1,46 +1,46 @@
 {
-  "name": "backend-skyfly-team01",
-  "version": "1.0.0",
-  "description": "Backend Final Project team 01 - SkyFly",
-  "main": "index.js",
-  "type": "commonjs",
-  "scripts": {
-    "dev": "nodemon index.js",
-    "start": "node index.js",
-    "db-push": "npx prisma db push",
-    "migrate": "npx prisma migrate dev",
-    "migrate-refresh": "npx prisma migrate reset",
-    "vercel-build": "prisma generate && prisma migrate deploy",
-    "db-seed": "node ./prisma/seeder/seederUser"
-  },
-  "keywords": [],
-  "author": "Yoga Andhika Rafi Viery Faris Naufal Lowis Falah",
-  "license": "ISC",
-  "devDependencies": {
-    "morgan": "^1.10.0",
-    "nodemon": "^3.1.0"
-  },
-  "dependencies": {
-    "@prisma/client": "^5.13.0",
-    "@supabase/supabase-js": "^2.43.2",
-    "bcrypt": "^5.1.1",
-    "cors": "^2.8.5",
-    "crypto": "^1.0.1",
-    "dotenv": "^16.4.5",
-    "ejs": "^3.1.10",
-    "express": "^4.19.2",
-    "googleapis": "^137.1.0",
-    "http-errors": "^2.0.0",
-    "joi": "^17.13.0",
-    "jsonwebtoken": "^9.0.2",
-    "multer": "^1.4.5-lts.1",
-    "nodemailer": "^6.9.13",
-    "nodemon": "^3.1.0",
-    "otpauth": "^9.2.4",
-    "pg": "^8.11.3",
-    "pg-hstore": "^2.3.4",
-    "prisma": "^5.13.0",
-    "swagger-jsdoc": "^6.2.8",
-    "swagger-ui-express": "^5.0.0"
-  }
+    "name": "backend-skyfly-team01",
+    "version": "1.0.0",
+    "description": "Backend Final Project team 01 - SkyFly",
+    "main": "index.js",
+    "type": "commonjs",
+    "scripts": {
+        "dev": "nodemon index.js",
+        "start": "node index.js",
+        "db-push": "npx prisma db push",
+        "migrate": "npx prisma migrate dev",
+        "migrate-refresh": "npx prisma migrate reset",
+        "vercel-build": "npx prisma generate && npx prisma migrate dev",
+        "db-seed": "node ./prisma/seeder/seederUser"
+    },
+    "keywords": [],
+    "author": "Yoga Andhika Rafi Viery Faris Naufal Lowis Falah",
+    "license": "ISC",
+    "devDependencies": {
+        "morgan": "^1.10.0",
+        "nodemon": "^3.1.0"
+    },
+    "dependencies": {
+        "@prisma/client": "^5.13.0",
+        "@supabase/supabase-js": "^2.43.2",
+        "bcrypt": "^5.1.1",
+        "cors": "^2.8.5",
+        "crypto": "^1.0.1",
+        "dotenv": "^16.4.5",
+        "ejs": "^3.1.10",
+        "express": "^4.19.2",
+        "googleapis": "^137.1.0",
+        "http-errors": "^2.0.0",
+        "joi": "^17.13.0",
+        "jsonwebtoken": "^9.0.2",
+        "multer": "^1.4.5-lts.1",
+        "nodemailer": "^6.9.13",
+        "nodemon": "^3.1.0",
+        "otpauth": "^9.2.4",
+        "pg": "^8.11.3",
+        "pg-hstore": "^2.3.4",
+        "prisma": "^5.13.0",
+        "swagger-jsdoc": "^6.2.8",
+        "swagger-ui-express": "^5.0.0"
+    }
 }
